pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "akari-cli"
version = "0.1.4"

authors = [
  { name="akariwill", email="mwildjrs23@example.com" },
]
description = "Akari Watanabe AI Assistant - Voice-first CLI & Web Assistant"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "websockets>=13.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    "rich>=14.3.3",
    "python-dotenv>=1.0.1",
]

[project.scripts]
akari = "akari_cli.akari_cli:main"

[tool.setuptools]
packages = ["akari_cli"]
include-package-data = true

[tool.setuptools.package-data]
"akari_cli" = ["data/**/*", "templates/**/*"]
